Spin off MissionManager class
