+++ src/bin/mission_manager/CMakeLists.txt
@@ -5,6 +5,7 @@ protobuf_generate_cpp(PROTO_SRCS PROTO_HDRS ${CMAKE_CURRENT_BINARY_DIR} config.p
 add_executable(${APP}
   app.cpp
   machine.cpp
+  MissionManager.cpp
   ${PROTO_SRCS} ${PROTO_HDRS})
 
 target_link_libraries(${APP}
